Fail gracefully when building library from source
Most users don't have the correct dependencies installed, so it's best
not to blow up when building from source doesn't work. The build error
is printed and the user is warned that they might need to manually
install the library.

Ideally, the build script should still be smart enough to install
everything auto-magically, but this will have to do for now.

Closes #14
Closes #18
Closes #21