Rewrite and DRY out callback macro
The macro now supports the ability to transform C parameters to a more
Julian form before passing them to the Julia function. Example:

    @callback Foo(x::Cint, str::Ptr{Cchar}) -> (x, bytestring(str))

The window value is no longer truncated when using anonymous callbacks.
It complicates the code and it's easy enough to use a dummy `_`
parameter when starting a new anonymous callback.

Munged global variables (via gensym) are now used for keeping track
of the Julia callback function. Window user pointer book-keeping was
removed.