Rewrite callback macro (#136)
The monolithic `callback` macro has been split into `callback`, `windowcallback`, and several other helper functions. The functionality is the same, but will hopefully be a little easier to maintain.

The `Window` type is now a thin wrapper over a handle again. The list of callback functions for a window is now independent. To prevent garbage collection, it lives inside a global `_window_callbacks` dictionary. Performance is retained via GLFW's user pointer feature.